pyproject.toml
[project]
name = "prueba-tecnica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-cache2>=0.2.2",
    "fastapi[standard]>=0.136.1",
    "httpx>=0.28.1",
    "logfire[fastapi]>=4.32.1",
    "pycountry>=26.2.16",
    "pydantic>=2.13.3",
    "pytest>=9.0.3",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.46.0",
]
